test(vscode-extension): extract shared helpers.ts — address review nits
Auto-review on #1963 flagged four nits collapsing to a single root
cause: `CONTRIBUTED_COMMANDS`, fixture-path resolution, the activation
`suiteSetup`, and the `assertCommandsStillRegistered` helper were
duplicated across the three integration-test files. Adding a new
contributed command would require updating four locations; adding a
new test suite would mean pasting the fixture + activation boilerplate
yet again.

Extract everything into `test/integration/helpers.ts`:

- `CONTRIBUTED_COMMANDS` (const)
- `fixture(name)`
- `activateExtension()` — opens `hello.cho`, activates the extension.
- `assertAllContributedCommandsRegistered()` — combines the existing
  `assertCommandsStillRegistered` + `getCommands(filterInternal=true)`
  assertion into a single named function.

Rewrite `activation.test.ts`, `preview.test.ts`, and
`lsp-config.test.ts` to import from the shared module. The old
per-file constants and activation blocks are removed. Behaviour is
unchanged — the tests exercise the same regression gates as before,
with fewer moving parts.

Closes #1964
